TradingAccountManager/requirements.txt
websockets
ccxt
discord
pyyaml